Loading & some visual changes

+++ app/src/main/java/com/example/android/mymovies/MainActivity.java
@@ -3,10 +3,13 @@ package com.example.android.mymovies;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.DisplayMetrics;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.CompoundButton;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -37,6 +40,7 @@ import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
     private Switch switchSort;
+    private ProgressBar progressBarLoading;
     private TextView textViewPopular, textViewTopRated;
     private RecyclerView recyclerView;
 
@@ -45,6 +49,10 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
 
     private static final int LOADER_ID = 133;
     private LoaderManager loaderManager;
+
+    private static int page = 1;
+    private static int methodSort;
+    private static boolean isLoading = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -59,6 +67,7 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
         switchSort.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                page = 1;
                 switchSorting(isChecked);
             }
         });
@@ -70,13 +79,18 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
             Intent intent = DetailsActivity.newInstance(MainActivity.this, movie.getId());
             startActivity(intent);
         });
-        adapter.setOnReachEndListener(() -> Toast.makeText(MainActivity.this,
-                "Конец списка", Toast.LENGTH_SHORT).show());
+        adapter.setOnReachEndListener(() -> {
+            if (!isLoading) {
+                downloadData(methodSort, page);
+            }
+        });
         LiveData<List<Movie>> moviesFromLD = viewModel.getMovies();
         moviesFromLD.observe(this, new Observer<List<Movie>>() {
             @Override
             public void onChanged(List<Movie> movies) {
-                adapter.setMovies(movies);
+                if (page == 1) {
+                    adapter.setMovies(movies);
+                }
             }
         });
     }
@@ -112,16 +126,16 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     }
 
     private void switchSorting(boolean isChecked) {
-        int methodSort = isChecked ? NetworkUtils.TOP_RATED : NetworkUtils.POPULARITY;
+        methodSort = isChecked ? NetworkUtils.TOP_RATED : NetworkUtils.POPULARITY;
         int popularColor = isChecked ? R.color.white : R.color.pink;
         int topRatedColor = isChecked ? R.color.pink : R.color.white;
         textViewPopular.setTextColor(ContextCompat.getColor(MainActivity.this, popularColor));
         textViewTopRated.setTextColor(ContextCompat.getColor(MainActivity.this, topRatedColor));
-        downloadData(methodSort, 1);
+        downloadData(methodSort, page);
     }
 
-    private void downloadData(int methodSort, int page) {
-        URL url = NetworkUtils.buildMovieUrl(methodSort, page);
+    private void downloadData(int methodOfSort, int page) {
+        URL url = NetworkUtils.buildMovieUrl(methodOfSort, page);
         Bundle bundle = new Bundle();
         bundle.putString("url", url.toString());
         loaderManager.restartLoader(LOADER_ID, bundle, this);
@@ -129,28 +143,50 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
 
     private void initViews() {
         switchSort = findViewById(R.id.switch_sort);
+        progressBarLoading = findViewById(R.id.pbar_loading);
         textViewPopular = findViewById(R.id.tv_most_popular);
         textViewTopRated = findViewById(R.id.tv_top_rated);
         recyclerView = findViewById(R.id.recycler_view_movies);
-        recyclerView.setLayoutManager(new GridLayoutManager(this, 2));
+        recyclerView.setLayoutManager(new GridLayoutManager(this, getColumnCount()));
+    }
+
+    private int getColumnCount() {
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        int width = (int) (displayMetrics.widthPixels / displayMetrics.density);
+        return width / 185 > 2 ? width / 185 : 2;
     }
 
     @NonNull
     @Override
     public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle bundle) {
         NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this, bundle);
+        jsonLoader.setOnStartLoadingListener(new NetworkUtils.JSONLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
         return jsonLoader;
     }
 
     @Override
     public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject data) {
         List<Movie> movies = JSONUtils.getMoviesFromJSON(data);
-        if (!movies.isEmpty()) {
-            viewModel.deleteAllMovies();
+        if (movies != null && !movies.isEmpty()) {
+            if (page == 1) {
+                viewModel.deleteAllMovies();
+                adapter.clear();
+            }
             for (Movie movie : movies) {
                 viewModel.insertMovie(movie);
             }
+            adapter.addMovies(movies);
+            page++;
         }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
         loaderManager.destroyLoader(LOADER_ID);
     }
